+++ src/Plainion.GraphViz/Model/Node.cs
@@ -6,17 +6,14 @@ namespace Plainion.GraphViz.Model
 {
     [Serializable]
     [DebuggerDisplay("{Id}")]
-    public class Node : IGraphItem
+    public class Node : AbstractGraphItem
     {
-        public Node( string id )
+        public Node(string id)
+            : base(id)
         {
-            Id = id;
-
             In = new List<Edge>();
             Out = new List<Edge>();
         }
-
-        public string Id { get; private set; }
 
         public IList<Edge> In { get; private set; }
         public IList<Edge> Out { get; private set; }